XP-1001 Rename ModuleService to ApplicationService
- Renamed ModuleService to ApplicationService
- Removed com.enonic.xp.app.ApplicationService (old one) and replaced the occurences with com.enonic.xp.module.ApplicationService (new one).
- Moved com.enonic.xp.module.ApplicationService to com.enonic.xp.app.ApplicationService
