+++ modules/core-api/src/main/java/com/enonic/xp/app/ApplicationService.java
@@ -1,10 +1,24 @@
 package com.enonic.xp.app;
 
+import com.google.common.annotations.Beta;
+
+import com.enonic.xp.module.Module;
+import com.enonic.xp.module.ModuleNotFoundException;
+import com.enonic.xp.module.Modules;
+
+@Beta
 public interface ApplicationService
 {
+    Module getModule( ApplicationKey key )
+        throws ModuleNotFoundException;
 
-    Applications getAll();
+    Modules getModules( ApplicationKeys keys );
 
-    Application getByKey( ApplicationKey applicationKey );
+    Modules getAllModules();
 
+    ClassLoader getClassLoader(Module module);
+
+    void startModule( ApplicationKey key );
+
+    void stopModule( ApplicationKey key );
 }
