+++ modules/core-impl/src/main/java/com/enonic/xp/core/impl/content/page/GetPageDescriptorsByModulesCommand.java
@@ -6,8 +6,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import com.enonic.xp.app.ApplicationKeys;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.module.Module;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.module.Modules;
 import com.enonic.xp.page.DescriptorKey;
 import com.enonic.xp.page.PageDescriptor;
@@ -25,14 +25,14 @@ final class GetPageDescriptorsByModulesCommand
 
     private ApplicationKeys applicationKeys;
 
-    private ModuleService moduleService;
+    private ApplicationService applicationService;
 
     private ResourceService resourceService;
 
 
     public PageDescriptors execute()
     {
-        final Modules modules = this.moduleService.getModules( this.applicationKeys );
+        final Modules modules = this.applicationService.getModules( this.applicationKeys );
         return getDescriptorsFromModules( modules );
     }
 
@@ -42,9 +42,9 @@ final class GetPageDescriptorsByModulesCommand
         return this;
     }
 
-    public final GetPageDescriptorsByModulesCommand moduleService( final ModuleService moduleService )
+    public final GetPageDescriptorsByModulesCommand applicationService( final ApplicationService applicationService )
     {
-        this.moduleService = moduleService;
+        this.applicationService = applicationService;
         return this;
     }
