+++ modules/core-impl/src/main/java/com/enonic/xp/core/impl/content/page/PageDescriptorServiceImpl.java
@@ -5,7 +5,7 @@ import org.osgi.service.component.annotations.Reference;
 
 import com.enonic.xp.app.ApplicationKey;
 import com.enonic.xp.app.ApplicationKeys;
-import com.enonic.xp.module.ModuleService;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.page.DescriptorKey;
 import com.enonic.xp.page.PageDescriptor;
 import com.enonic.xp.page.PageDescriptorService;
@@ -17,7 +17,7 @@ import com.enonic.xp.schema.mixin.MixinService;
 public final class PageDescriptorServiceImpl
     implements PageDescriptorService
 {
-    private ModuleService moduleService;
+    private ApplicationService applicationService;
 
     private MixinService mixinService;
 
@@ -35,7 +35,7 @@ public final class PageDescriptorServiceImpl
     public PageDescriptors getByModule( final ApplicationKey applicationKey )
     {
         return new GetPageDescriptorsByModuleCommand().
-            moduleService( this.moduleService ).
+            applicationService( this.applicationService ).
             mixinService( this.mixinService ).
             resourceService( this.resourceService ).
             applicationKey( applicationKey ).execute();
@@ -45,16 +45,16 @@ public final class PageDescriptorServiceImpl
     public PageDescriptors getByModules( final ApplicationKeys applicationKeys )
     {
         return new GetPageDescriptorsByModulesCommand().
-            moduleService( this.moduleService ).
+            applicationService( this.applicationService ).
             mixinService( this.mixinService ).
             resourceService( this.resourceService ).
             applicationKeys( applicationKeys ).execute();
     }
 
     @Reference
-    public void setModuleService( final ModuleService moduleService )
+    public void setApplicationService( final ApplicationService applicationService )
     {
-        this.moduleService = moduleService;
+        this.applicationService = applicationService;
     }
 
     @Reference
