+++ modules/core-impl/src/main/java/com/enonic/xp/core/impl/content/page/region/GetLayoutDescriptorsByModulesCommand.java
@@ -11,7 +11,7 @@ final class GetLayoutDescriptorsByModulesCommand
 
     public LayoutDescriptors execute()
     {
-        final Modules modules = this.moduleService.getModules( this.applicationKeys );
+        final Modules modules = this.applicationService.getModules( this.applicationKeys );
         return getDescriptorsFromModules( modules );
     }
