+++ modules/core-impl/src/main/java/com/enonic/xp/core/impl/content/page/region/GetPartDescriptorsByModuleCommand.java
@@ -11,7 +11,7 @@ final class GetPartDescriptorsByModuleCommand
 
     public PartDescriptors execute()
     {
-        final Module module = this.moduleService.getModule( this.applicationKey );
+        final Module module = this.applicationService.getModule( this.applicationKey );
         return getDescriptorsFromModule( module );
     }
