+++ modules/core-impl/src/test/java/com/enonic/xp/core/impl/app/ModuleRegistryImplTest.java
@@ -1,4 +1,4 @@
-package com.enonic.xp.core.impl.module;
+package com.enonic.xp.core.impl.app;
 
 import java.net.URL;
 import java.net.URLClassLoader;
