+++ modules/core-impl/src/test/resources/bundles/bundle1/app/site.xml
