+++ modules/core-impl/src/test/resources/bundles/bundle3/app/site.xml
