+++ modules/core-module/build.gradle
@@ -3,7 +3,6 @@ apply from: "$rootDir/gradle/osgi.gradle"
 dependencies {
     compile project( ':core-api' )
     testCompile project( path: ':core-api', configuration: 'testOutput' )
-    testCompile 'com.googlecode.pojosr:de.kalpatec.pojosr.framework:0.2.1'
 }
 
 bundle {
