+++ modules/core-schema/src/main/java/com/enonic/xp/core/impl/schema/relationship/RelationshipTypeServiceImpl.java
@@ -16,8 +16,8 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 
 import com.enonic.xp.app.ApplicationKey;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.module.Module;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.schema.relationship.RelationshipType;
 import com.enonic.xp.schema.relationship.RelationshipTypeName;
 import com.enonic.xp.schema.relationship.RelationshipTypeService;
@@ -29,7 +29,7 @@ public final class RelationshipTypeServiceImpl
 {
     private final Map<ApplicationKey, RelationshipTypes> relationshipTypesMap;
 
-    private ModuleService moduleService;
+    private ApplicationService applicationService;
 
     private BundleContext context;
 
@@ -68,7 +68,7 @@ public final class RelationshipTypeServiceImpl
         relationshipTypeList.addAll( systemRelationshipTypes.getList() );
 
         //Gets for each module the RelationshipTypes
-        for ( Module module : this.moduleService.getAllModules() )
+        for ( Module module : this.applicationService.getAllModules() )
         {
             final RelationshipTypes relationshipTypes = getByModule( module.getKey() );
             relationshipTypeList.addAll( relationshipTypes.getList() );
@@ -97,7 +97,7 @@ public final class RelationshipTypeServiceImpl
         else
         {
             //Else, loads the corresponding bundle relation types
-            final Module module = this.moduleService.getModule( applicationKey );
+            final Module module = this.applicationService.getModule( applicationKey );
             if ( module != null )
             {
                 final BundleRelationshipTypeLoader bundleRelationshipTypeLoader = new BundleRelationshipTypeLoader( module.getBundle() );
@@ -114,9 +114,9 @@ public final class RelationshipTypeServiceImpl
     }
 
     @Reference
-    public void setModuleService( final ModuleService moduleService )
+    public void setApplicationService( final ApplicationService applicationService )
     {
-        this.moduleService = moduleService;
+        this.applicationService = applicationService;
     }
 
     @Override
