+++ modules/core-schema/src/test/java/com/enonic/xp/core/impl/schema/content/ContentTypeRegistryImplTest.java
@@ -8,9 +8,9 @@ import org.osgi.framework.BundleEvent;
 import org.osgi.service.component.ComponentContext;
 
 import com.enonic.xp.app.ApplicationKey;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.core.impl.schema.AbstractBundleTest;
 import com.enonic.xp.module.Module;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.module.Modules;
 import com.enonic.xp.schema.content.ContentType;
 import com.enonic.xp.schema.content.ContentTypeName;
@@ -30,7 +30,7 @@ public class ContentTypeRegistryImplTest
 
     private Module myModule;
 
-    private ModuleService moduleService;
+    private ApplicationService applicationService;
 
     private ContentTypeRegistryImpl service;
 
@@ -50,7 +50,7 @@ public class ContentTypeRegistryImplTest
         Mockito.when( myModule.getBundle() ).thenReturn( myBundle );
 
         //Mocks the module service
-        moduleService = Mockito.mock( ModuleService.class );
+        applicationService = Mockito.mock( ApplicationService.class );
 
         //Mocks the ComponentContext
         final ComponentContext componentContext = Mockito.mock( ComponentContext.class );
@@ -58,7 +58,7 @@ public class ContentTypeRegistryImplTest
 
         //Creates the service to test
         service = new ContentTypeRegistryImpl();
-        service.setModuleService( moduleService );
+        service.setApplicationService( applicationService );
 
         //Starts the service
         service.start( componentContext );
@@ -67,7 +67,7 @@ public class ContentTypeRegistryImplTest
     @Test
     public void test_empty()
     {
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
 
         ContentTypes contentTypes = service.getAll();
         assertNotNull( contentTypes );
@@ -86,8 +86,8 @@ public class ContentTypeRegistryImplTest
     {
 
         Modules modules = Modules.from( myModule );
-        Mockito.when( moduleService.getAllModules() ).thenReturn( modules );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( myModule );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( modules );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( myModule );
 
         ContentTypes contentTypes = service.getAll();
         assertNotNull( contentTypes );
@@ -100,8 +100,8 @@ public class ContentTypeRegistryImplTest
         ContentType contentType = service.get( this.myContentType.getName() );
         assertNotNull( contentType );
 
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( null );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( null );
         service.bundleChanged( new BundleEvent( BundleEvent.UNINSTALLED, myBundle ) );
 
         test_empty();
@@ -111,7 +111,7 @@ public class ContentTypeRegistryImplTest
     public void test_get_system_module()
     {
 
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
 
         ContentTypes contentTypes = service.getAll();
         assertNotNull( contentTypes );
