+++ modules/core-schema/src/test/java/com/enonic/xp/core/impl/schema/mixin/MixinServiceImplTest.java
@@ -8,9 +8,9 @@ import org.osgi.framework.BundleEvent;
 import org.osgi.service.component.ComponentContext;
 
 import com.enonic.xp.app.ApplicationKey;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.core.impl.schema.AbstractBundleTest;
 import com.enonic.xp.module.Module;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.module.Modules;
 import com.enonic.xp.schema.content.ContentType;
 import com.enonic.xp.schema.content.ContentTypeName;
@@ -32,7 +32,7 @@ public class MixinServiceImplTest
 
     private Module myModule;
 
-    private ModuleService moduleService;
+    private ApplicationService applicationService;
 
     private MixinServiceImpl service;
 
@@ -53,7 +53,7 @@ public class MixinServiceImplTest
         Mockito.when( myModule.getBundle() ).thenReturn( myBundle );
 
         //Mocks the module service
-        moduleService = Mockito.mock( ModuleService.class );
+        applicationService = Mockito.mock( ApplicationService.class );
 
         //Mocks the ComponentContext
         final ComponentContext componentContext = Mockito.mock( ComponentContext.class );
@@ -61,7 +61,7 @@ public class MixinServiceImplTest
 
         //Creates the service to test
         service = new MixinServiceImpl();
-        service.setModuleService( moduleService );
+        service.setApplicationService( applicationService );
 
         //Starts the service
         service.start( componentContext );
@@ -70,7 +70,7 @@ public class MixinServiceImplTest
     @Test
     public void test_empty()
     {
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
 
         Mixins mixins = service.getAll();
         assertNotNull( mixins );
@@ -89,8 +89,8 @@ public class MixinServiceImplTest
     {
 
         Modules modules = Modules.from( myModule );
-        Mockito.when( moduleService.getAllModules() ).thenReturn( modules );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( myModule );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( modules );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( myModule );
 
         Mixin mixin = service.getByLocalName( "mixin1" );
         assertNotNull( mixin );
@@ -102,8 +102,8 @@ public class MixinServiceImplTest
     {
 
         Modules modules = Modules.from( myModule );
-        Mockito.when( moduleService.getAllModules() ).thenReturn( modules );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( myModule );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( modules );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( myModule );
 
         ContentType contentType = ContentType.create().
             superType( ContentTypeName.structured() ).
@@ -122,8 +122,8 @@ public class MixinServiceImplTest
     {
 
         Modules modules = Modules.from( myModule );
-        Mockito.when( moduleService.getAllModules() ).thenReturn( modules );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( myModule );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( modules );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( myModule );
 
         Mixins mixins = service.getAll();
         assertNotNull( mixins );
@@ -136,8 +136,8 @@ public class MixinServiceImplTest
         Mixin mixin = service.getByName( this.mixin1.getName() );
         assertNotNull( mixin );
 
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
-        Mockito.when( moduleService.getModule( myApplicationKey ) ).thenReturn( null );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getModule( myApplicationKey ) ).thenReturn( null );
         service.bundleChanged( new BundleEvent( BundleEvent.UNINSTALLED, myBundle ) );
 
         test_empty();
@@ -147,7 +147,7 @@ public class MixinServiceImplTest
     public void test_get_system_module()
     {
 
-        Mockito.when( moduleService.getAllModules() ).thenReturn( Modules.empty() );
+        Mockito.when( applicationService.getAllModules() ).thenReturn( Modules.empty() );
 
         Mixins mixins = service.getAll();
         assertNotNull( mixins );
