+++ modules/portal-impl/src/main/java/com/enonic/xp/portal/impl/services/PortalServices.java
@@ -1,9 +1,9 @@
 package com.enonic.xp.portal.impl.services;
 
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.content.ContentService;
 import com.enonic.xp.image.ImageFilterBuilder;
 import com.enonic.xp.image.ImageScaleFunctionBuilder;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.page.PageDescriptorService;
 import com.enonic.xp.page.PageTemplateService;
 import com.enonic.xp.portal.impl.controller.ControllerScriptFactory;
@@ -14,7 +14,7 @@ import com.enonic.xp.site.SiteService;
 
 public interface PortalServices
 {
-    ModuleService getModuleService();
+    ApplicationService getApplicationService();
 
     ControllerScriptFactory getControllerScriptFactory();
