+++ modules/portal-impl/src/test/java/com/enonic/xp/portal/impl/controller/AbstractControllerTest.java
@@ -15,8 +15,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
 import com.enonic.xp.app.ApplicationKey;
+import com.enonic.xp.app.ApplicationService;
 import com.enonic.xp.module.Module;
-import com.enonic.xp.module.ModuleService;
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
 import com.enonic.xp.portal.impl.postprocess.PostProcessorImpl;
@@ -63,12 +63,12 @@ public abstract class AbstractControllerTest
         final Module module = Mockito.mock( Module.class );
         Mockito.when( module.getBundle() ).thenReturn( bundle );
 
-        final ModuleService moduleService = Mockito.mock( ModuleService.class );
-        Mockito.when( moduleService.getModule( ApplicationKey.from( "mymodule" ) ) ).thenReturn( module );
-        Mockito.when( moduleService.getClassLoader( Mockito.any() ) ).thenReturn( getClass().getClassLoader() );
+        final ApplicationService applicationService = Mockito.mock( ApplicationService.class );
+        Mockito.when( applicationService.getModule( ApplicationKey.from( "mymodule" ) ) ).thenReturn( module );
+        Mockito.when( applicationService.getClassLoader( Mockito.any() ) ).thenReturn( getClass().getClassLoader() );
 
         final ScriptServiceImpl scriptService = new ScriptServiceImpl();
-        scriptService.setModuleService( moduleService );
+        scriptService.setApplicationService( applicationService );
 
         this.factory = new ControllerScriptFactoryImpl();
         this.factory.setScriptService( scriptService );
