+++ modules/portal-impl/src/test/java/com/enonic/xp/portal/impl/resource/asset/AssetResourceTest.java
@@ -15,7 +15,7 @@ public class AssetResourceTest
     protected void configure()
         throws Exception
     {
-        configureModuleService();
+        configureApplicationService();
     }
 
     @Test
